Split CI testing into separate actor and non-actor stages
This feels a little repetitive and messy but will allow us to run the required RUSTFLAGS for coverage report generation on most of the code while avoiding conflict because WASM doesn't support profiling builds it requires. Support for that is apparently an LLVM issue rather than a Rust one but for now it's easier to run things separately.