Method match macro and example actor (#52)
* Add match_method macro to generate method dispatch matches

* Add match_method rule to handle the default case

Instead of forcing `_ => None`, this lets the user decide how to handle things

* Add unit tests

* Add extra rules to handle blocks with/without commas

Without this, the macro required a comma at the end of each match item and weird errors resulted if it was missed

* Add basic greeter example

* Don't check all targets and features during CI

This can trigger some messy crate version mismatch issues with fvm_integration tests

* Use same nightly version override as ref-fvm repo does

Recently nightly builds removed `backtrace()` from `Error` which the `fvm` crate is using. See https://github.com/rust-lang/rust/commit/b2bbca393384c6d7e343decef5a927e16f003038 for the commit behind it but we can work around it for now by using an older nightly version. I've set it to match the version used by ref-fvm to hopefully avoid any other compatibility issues with the many crates we use from that repo.

* Add a README to the greeter example

* Split CI testing into separate actor and non-actor stages

This feels a little repetitive and messy but will allow us to run the required RUSTFLAGS for coverage report generation on most of the code while avoiding conflict because WASM doesn't support profiling builds it requires. Support for that is apparently an LLVM issue rather than a Rust one but for now it's easier to run things separately.

* Remove out-of-place comment