Receiver hook call guard for mint/transfer (#82)
* Add a callable ReceiverHookGuard to manage calling receiver hook after minting

It holds the parameters for the receiver hook call as well as the original return data from the minting call (currently a TokenAmount representing total supply). If not called, it will panic when dropped out of scope, effectively aborting the (half-finished) transaction.

* Use ReceiverHookGuard for transfer operations

* Fix tests

* Fix ignored tests

A couple of them have changed, removing post-failure balance checks because the state is not reverted after hook call failure. They still check for the correct error result, but the actor-level behaviour is to abort rather than manually revert the state now.

* Remove return value from ReceiverHookGuard and return it directly from the originating function

* Keep ownership of TokensReceivedParams instead of moving it in and out of an Option

* Include address and params in 'receiver hook not called' panic message

* Rename ReceiverHookGuard to just ReceiverHook

* Add comments explaining ReceiverHook

* Add a test-only Token::replace function to revert state in unit tests that need it

* Add unit tests for ReceiverHook