Add SharedMemoryBlockstore for use in ActorRuntime
This is a reference-counted wrapper around MemoryBlockstore. Cloning a SharedMemoryBlockstore will reference the same underlying MemoryBlockstore instance. This is useful in some unit testing cases where multiple components require ownership of the ActorRuntime, but expect a single Blockstore. Behaviour more closely matches the native Blockstore wrapper (which is an empty struct wrapping the syscalls, so every instance operates on the one underlying store)