Use a reference to the ActorRuntime in FRC46 Token handle instead of owning it
The Token handle already held a mutable reference to the TokenState, I've changed it to handle ActorRuntime the same way. The runtime is almost always created alongside it so impact on existing code has been minimal. This solves complications I had with the factory token implementation and removes the need to clone the runtime there.