FRC46 factory token unit test coverage, ActorRuntime improvements (#190)
* Add Clone trait bound to ActorRuntime

The Blockchain and Syscalls parts implement Clone but ActorRuntime::clone can't be called without the trait bound being present everywhere

* Add caller() function to Syscalls

This currently returns a fixed ActorID of 1 for the FakeSyscalls implementation. It's only used in unit tests so it may require changes later to add a way to set it, if tests can't be built around that assumption.

* Use ActorRuntime in FactoryToken

This is in preparation for adding unit tests to the token_impl code as that will require fake syscalls and the memory blockstore

* Update FactoryToken integration tests

* Add unit tests for FactoryToken minting

* Add set_root to ActorRuntime

* Get caller ActorID/Address from the ActorRuntime instead of syscall

* Add SharedMemoryBlockstore for use in ActorRuntime

This is a reference-counted wrapper around MemoryBlockstore. Cloning a SharedMemoryBlockstore will reference the same underlying MemoryBlockstore instance. This is useful in some unit testing cases where multiple components require ownership of the ActorRuntime, but expect a single Blockstore. Behaviour more closely matches the native Blockstore wrapper (which is an empty struct wrapping the syscalls, so every instance operates on the one underlying store)

* Add comment explaining how cloning the runtime can lead to split blockstores if using MemoryBlockstore

* Add tests for disabling minting on FactoryToken

* Make caller ID configurable in FakeSyscalls

* Add more unit tests for FactoryToken

* Fix test formatting

* Add even more unit tests to FactoryToken

* Use a reference to the ActorRuntime in FRC46 Token handle instead of owning it

The Token handle already held a mutable reference to the TokenState, I've changed it to handle ActorRuntime the same way. The runtime is almost always created alongside it so impact on existing code has been minimal. This solves complications I had with the factory token implementation and removes the need to clone the runtime there.

* Revert "Add Clone trait bound to ActorRuntime"

This reverts commit c987b0e9d77bbdcb251fa452cb1b108d27bd8387.

* cleanup: remove TODOs

* update commends in FakeSyscalls

* remove last Clone remnants

* pass an ActorRuntime into construct_token

* update for new fvm/crate versions

* Add ActorError to ExitCode conversion to simplify error handling for ActorRuntime users

* minor cleanup

* Replace more get/set root cid syscalls with ActorRuntime methods