test: fetch actors bundle from github instead of using cargo
Primarily, this removes an annoying circular dependency between these
utilities and the actors. Additionally, this makes it very clear that
the actors we're building won't use the code in this repo. Unlike normal
rust builds, actor bundle builds don't automatically patch workspaces.

I've configured it to fetch v15 and we'll have to update that as we go.
If you want to test with newer actors, you'll need to build a bundle
yourself and drop it in `testing/bundles/builtin-actors.car`.

Part of https://github.com/filecoin-project/builtin-actors/issues/1664